test(logic): split combat resolver deployment and general medal cases
Extract deployment-limit-with-nationalism and general-medal progression scenarios from the oversized part1 combat resolver test file so each test file stays under the logic test size gate while preserving coverage for issue #2288.
